requirements.txt
streamlit
langchain
langchain-community
supabase
python-dotenv
sentence-transformers
pypdf
openai